[New] Added test cases for pages admin roles list
