+++ app/javascript/components/action-button/component.jsx
@@ -41,6 +41,7 @@ function Component({
 
   return (
     <ButtonType
+    data-testid="action-button"
       id={buttonID}
       icon={icon}
       cancel={cancel}
